remove deprecated, update deps, refactor
