+++ .gitignore
@@ -5,3 +5,4 @@ node_modules
 package-lock.json
 shrinkwrap.yaml
 yarn.lock
+pnpm-lock.yaml
