--- .npmignore
@@ -1,10 +1,0 @@
-/.vscode
-/src
-/test
-/coverage
-.gitignore
-node_modules
-package-lock.json
-shrinkwrap.yaml
-yarn.lock
-tsconfig.json
